WIP : add handling of input when process is running
