+++ gridcapa-task-manager-app/src/main/java/com/farao_community/farao/gridcapa/task_manager/app/repository/ProcessFileRepository.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.farao_community.farao.gridcapa.task_manager.app;
+package com.farao_community.farao.gridcapa.task_manager.app.repository;
 
 import com.farao_community.farao.gridcapa.task_manager.app.entities.ProcessFile;
 import org.springframework.data.jpa.repository.JpaRepository;
