+++ gridcapa-task-manager-app/src/main/java/com/farao_community/farao/gridcapa/task_manager/app/service/FileEventType.java
@@ -1,0 +1,7 @@
+package com.farao_community.farao.gridcapa.task_manager.app.service;
+
+public enum FileEventType {
+    AVAILABLE,
+    UPDATED,
+    DELETED
+}
