+++ gridcapa-task-manager-app/src/test/java/com/farao_community/farao/gridcapa/task_manager/app/TaskManagerTestUtil.java
@@ -6,6 +6,7 @@
  */
 package com.farao_community.farao.gridcapa.task_manager.app;
 
+import com.farao_community.farao.gridcapa.task_manager.app.service.MinioHandler;
 import com.farao_community.farao.minio_adapter.starter.MinioAdapter;
 import io.minio.messages.Event;
 import org.mockito.Mockito;
@@ -23,10 +24,10 @@ public final class TaskManagerTestUtil {
     public static Event createEvent(MinioAdapter minioAdapter, String processTag, String fileGroup, String fileType, String fileKey, String validityInterval, String fileUrl) {
         Event event = Mockito.mock(Event.class);
         Map<String, String> metadata = Map.of(
-            TaskManager.FILE_GROUP_METADATA_KEY, fileGroup,
-            TaskManager.FILE_TARGET_PROCESS_METADATA_KEY, processTag,
-            TaskManager.FILE_TYPE_METADATA_KEY, fileType,
-            TaskManager.FILE_VALIDITY_INTERVAL_METADATA_KEY, validityInterval
+                MinioHandler.FILE_GROUP_METADATA_KEY, fileGroup,
+                MinioHandler.FILE_TARGET_PROCESS_METADATA_KEY, processTag,
+                MinioHandler.FILE_TYPE_METADATA_KEY, fileType,
+                MinioHandler.FILE_VALIDITY_INTERVAL_METADATA_KEY, validityInterval
         );
         //The following mock is not use in all test that call this methods. With the "lenient" add you avoid an exception
         Mockito.lenient().when(event.userMetadata()).thenReturn(metadata);
